feat: 전역 변수 관리
