+++ src/entities/post/models/index.ts
@@ -1,3 +1,1 @@
-export { PostContext, usePostContext } from "./use-post-context";
-
-export type { NewPost, Post, PostWithAuthor } from "./post.types";
+export type { NewPost, Post, Tag } from "./post.types";
